third_party/blink/web_tests/external/wpt/service-workers/service-worker/clients-matchall-blob-url-worker.https-expected.txt

This is a testharness.js-based test.
[FAIL] Test Clients.matchAll() with a blob URL worker client.
  assert_equals: expected 1 but got 0
[FAIL] Test Clients.matchAll() with an uncontrolled blob URL worker client.
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
